Implement help text DSL for commands
This adds a new DSL for defining help text on commands.  The `summary`,
`description`, and `examples` commands at the class level attach parts of the
documentation to the command object itself.

While these presently allow us to attach the documentation and provide no
mechanism to communicate it to the user, they are the basis for building that
facility and allow us to work in parallel on building and presenting the docs.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>