Razor version string must not include a newline
An undetected bug meant that some razor versions could include a newline when
generating from git metadata.  This hadn't caused any practical problems until
we started using that as part of the etag for cache control.

This fixes the issue, and adds a test to try and catch this if it crops up
again somewhere.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>