Add user help to all commands
This adds user-facing help to all commands, based on the existing
documentation.  This will be automatically exposed to the client through the
new command metadata API, and will permit this information to be shared with
the client automatically -- as well as future updates to help about the
commands as we improve the software.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>