Automatically type-check array and object entities in validation
When an `object` or `array` helper is used for an attribute, we already know
what type the entity is -- so we can automatically set that check in the
attribute code.

This allows things like `size` checks to be applied more naturally to
`object` or `array` attributes, too, without having to manually add the
`type` check.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>