Add shared example specs for required attributes
It is a common case when testing commands that there is a list of minimum required attributes in the database validation which can differ from the command validation. This adds shared examples which can programmatically test that the minimum case works, then that removing any one attribute fails.

Note that this will not catch cases where the command validation is stronger than the database validation.