Migrate all object creation to a new API
The semantics offered by the Sequel ORM are fairly traditional: object
creation or searching is possible, but it will simply raise an error when
items that violate unique constraints are found.

We want to move beyond that, to a world where creation is idempotent: if the
object logically exists in the database, but is identical to what you
requested, then we should treat this as a success rather than an error.

This moves all the object creation to a new API, "import", on our data
objects.  This allows us to add the semantics without overriding the core
behaviour of the Sequel ORM.

https://tickets.puppetlabs.com/browse/RAZOR-185

Signed-off-by: Daniel Pittman <daniel@rimspace.net>