Put `node-metadata` back in `create-policy`
The create-policy command should accept the node-metadata attribute to allow the policy to seed metadata onto nodes that bind to it. This was previously implemented but the parameter on the create-policy command was lost in the move to the new validation framework. This just adds it back in.

Based on work by Chris Portman.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-207