Make sure that checkin does not try to bind a node that is installed
We used to solely consider during checkin whether a node had a policy
assigned to it to determine whether we should try and bind a node to a new
policy. That was wrong, as it is possible to have nodes that are installed
but have no policy. To protect against inadvertant data loss in the case
such a node goes through checkin, we now make sure that nodes that either
have a policy assigned or are marked as installed are never bound to a new
policy.

Issue: https://tickets.puppetlabs.com/browse/RAZOR-263