When 'protect_new_nodes' is true, make sure we boot new nodes into the MK
Even when protect_new_nodes is true we should run new nodes through the
microkernel once so we have facts for them. After that, we will boot such
nodes locally.

This change also adds a lot of (but not enough) tests around the
functionality connected to booting and checkin of nodes in various
combinations of bound/installed/registered.

Issue: https://tickets.puppetlabs.com/browse/RAZOR-264